pyproject.toml
##############################################
# The MIT License (MIT)
# Copyright (c) 2025 Kevin Walchko
# see LICENSE for full details
##############################################

[tool.poetry]
name = "pyterm"
version = "2025.04.20"
description = "serial terminal"
authors = ["walchko <walchko@users.noreply.github.com>"]
readme = "readme.md"
license = "MIT"
homepage = "https://pypi.org/project/pyterm/"
repository = "https://github.com/MomsFriendlyRobotCompany/pyterm"
# documentation = "http://..."
keywords = ["serial", "terminal"]
classifiers = [
    'Development Status :: 4 - Beta',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Topic :: Software Development :: Libraries',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Topic :: Software Development :: Libraries :: Application Frameworks'
]

[tool.poetry.dependencies]
python = ">=3.10"
colorama = "*"
pyserial = "*"

[tool.poetry.dev-dependencies]
pytest = "*"

[tool.poetry.scripts]
pyterm = "pyterm.terminal:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
